services/ingest-outage/requirements.txt
requests
psycopg2-binary
apscheduler
pandas
lxml
